feat: Add retry logic for fetching situation in cloud function
The code changes introduce retry logic for fetching the situation in the cloud function "fetchSituationForCreatingDialog". The function will now attempt to fetch the situation multiple times, up to a maximum of 3 retries, in case of any errors. If the maximum number of retries is exceeded, an error message will be logged. This enhancement improves the reliability of the function.